added del installed gateways functionality
