+++ ui/src/index.css
@@ -1,6 +1,10 @@
 @font-face {
-    font-family: "urbitSans";
+    /* font-family: "urbitSans";
     src: url("./fonts/UrbitSansWeb-Regular.woff2") format('woff2');
+    10:09 */
+    font-family: "urbitSans";
+    src: url('https://media.urbit.org/fonts/UrbitSans/UrbitSansVFWeb-Regular.woff2')
+       format('woff2');
     /* font-family: "urbitSansMono";
     src: url("./fonts/UrbitSansMonoWeb-Thin.woff2") format('woff2'); */
   }
